Implemented Oxygen System
Changes:
- Updated .gitignore to ignore the Riderlink plugin and the .idea folder. Prevents updates in the repo due to simply coding on a different machine, for example.
- As in title, implemented oxygen system. The mechanisms are currently in place and it should also have some hooks on it so that it can eventually be connected with appropriate UI elements.

Information:
The logic for the Oxygen System is currently primarily held within BP_TopDownCharacter with the trigger being an instance of BP_OxygenZoneTrigger. If we wanted to make this a bit more flexible in the game we could shift the functionality into the player controller instead.

This would allow the player to have one consistent oxygen value instead of a unique value for each instance of the BP_TopDownCharacter (let me know if this does not make sense :sweat_smile:

I would appreciate any testing that could be done just to ensure it's as robust as possible.