Zoning Tile Indicators & Area Calculation
Today I created the Indicators using a Niagara Beam emitter which is updated when the player moves around the cursor. This allows the player to understand where the area will be created, Holding Left Shift key allows for more Posts to be created, increasing the area size. I also created the Area calculation which will be used to determine the crop yield when plants are fully grown.